Refactor: auth, user 분리 (#15)
